requirements.txt
argon2-cffi==21.1.0
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
cloudinary==1.26.0
dj-database-url==0.5.0
Django==3.2.7
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2==2.9.1
pycparser==2.20
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
urllib3==1.26.6
whitenoise==5.3.0
